+++ tests/lib.rs
@@ -119,6 +119,19 @@ fn test_tile_property() {
 }
 
 #[test]
+fn test_layer_property() {
+    let r = read_from_file(&Path::new("assets/tiled_base64.tmx")).unwrap();
+    let prop_value: String = if let Some(&PropertyValue::StringValue(ref v)) =
+        r.layers[0].properties.get("prop3")
+    {
+        v.clone()
+    } else {
+        String::new()
+    };
+    assert_eq!("Line 1\r\nLine 2\r\nLine 3,\r\n  etc\r\n   ", prop_value);
+}
+
+#[test]
 fn test_object_group_property() {
     let r = parse_map_without_source("assets/tiled_object_groups.tmx").unwrap();
     let prop_value: bool = if let Some(&PropertyValue::BoolValue(ref v)) = r.object_groups[0]
@@ -184,4 +197,17 @@ fn test_ldk_export() {
     } else {
         assert!(false, "It is wrongly recognised as an infinite map");
     }
+}
+
+#[test]
+fn test_object_property() {
+    let r = read_from_file(&Path::new("assets/tiled_object_property.tmx")).unwrap();
+    let prop_value = if let Some(PropertyValue::ObjectValue(v)) =
+        r.object_groups[0].objects[0].properties.get("object property")
+    {
+        *v
+    } else {
+        0
+    };
+    assert_eq!(3, prop_value);
 }